+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### v0.2.6
 
  * New feature: custom password preference with confirmation (Issue #26)
+ * New feature: use an individual password or PIN to lock the app (Issue #23)
  * New feature: support for Panic Trigger (PR #27 by carmebar)
  * Bug fix: OpenPGP with security token (Issue #20, PR #25 by carmebar)
  * Style/UI: add Contributors, Translators and Translate to About